Don't update page on visibility change if there's a query
